requirements.txt
-e git+https://github.com/django/djnago.git@1.7b1#egg=Django
mock
requests
stripe
psycopg2
djangorestframework
django-embed-video
